refactor to jest
